+++ packages/datadog_flutter_plugin/test/rum/rum_long_task_observer_test.dart
@@ -1,6 +1,7 @@
 // Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019-Present Datadog, Inc.
+@TestOn('vm')
 
 import 'dart:async';
 
@@ -140,15 +141,15 @@ void main() {
 
     observer.didChangeAppLifecycleState(AppLifecycleState.inactive);
     await tester.runAsync(() async {
-      await Future<void>.delayed(const Duration(milliseconds: 100));
+      await Future<void>.delayed(const Duration(milliseconds: 150));
     });
-    await tester.pump(const Duration(milliseconds: 200));
+    await tester.pump(const Duration(milliseconds: 300));
 
     observer.didChangeAppLifecycleState(AppLifecycleState.resumed);
     await tester.runAsync(() async {
-      await Future<void>.delayed(const Duration(milliseconds: 100));
+      await Future<void>.delayed(const Duration(milliseconds: 150));
     });
-    await tester.pump(const Duration(milliseconds: 200));
+    await tester.pump(const Duration(milliseconds: 300));
 
     verify(() => mockRum.reportLongTask(any())).called(1);