+++ packages/datadog_session_replay/lib/src/capture/element_recorders/container_recorder.dart
@@ -16,10 +16,13 @@ class ContainerRecorder implements ElementRecorder {
   const ContainerRecorder(this.keyGenerator);
 
   @override
+  List<Type> get handlesTypes => [ColoredBox, Material, DecoratedBox];
+
+  @override
   CaptureNodeSemantics? captureSemantics(
     Element element,
     CapturedViewAttributes attributes,
-    CapturePrivacy capturePrivacy,
+    TreeCapturePrivacy capturePrivacy,
   ) {
     final widget = element.widget;
     // Material is also considered a container
@@ -43,6 +46,8 @@ class ContainerRecorder implements ElementRecorder {
         style ??= ContainerStyle(backgroundColor: null);
         break;
     }
+
+    attributes = _adjustAttributesForShape(widget, attributes);
 
     final key = keyGenerator.keyForElement(element);
     final node = ContainerNode(attributes, wireframeId: key, style: style!);
@@ -76,5 +81,37 @@ class ContainerRecorder implements ElementRecorder {
       borderWidth: borderStyle?.width,
       cornerRadius: borderStyle?.cornerRadius ?? 0.0,
     );
+  }
+
+  CapturedViewAttributes _adjustAttributesForShape(
+    Widget widget,
+    CapturedViewAttributes attributes,
+  ) {
+    CircleBorder? circleBorder;
+    switch (widget) {
+      case Material widget:
+        if (widget.shape case final CircleBorder circle) {
+          circleBorder = circle;
+        }
+        break;
+      case DecoratedBox box:
+        final decoration = box.decoration;
+        if (decoration is ShapeDecoration && decoration.shape is CircleBorder) {
+          circleBorder = decoration.shape as CircleBorder;
+        }
+        break;
+    }
+    if (circleBorder != null) {
+      // Need to adjust position, width, and height so that this actually appears as
+      // a circle in Session Replay
+      final center = attributes.paintBounds.center;
+      final shortSide = attributes.paintBounds.shortestSide;
+      attributes = CapturedViewAttributes(
+        paintBounds: Rect.fromCircle(center: center, radius: shortSide / 2),
+        scaleX: attributes.scaleX,
+        scaleY: attributes.scaleY,
+      );
+    }
+    return attributes;
   }
 }