+++ packages/datadog_session_replay/lib/src/capture/element_recorders/custom_paint_recorder.dart
@@ -18,10 +18,13 @@ class CustomPaintRecorder implements ElementRecorder {
   const CustomPaintRecorder(this.keyGenerator);
 
   @override
+  List<Type> get handlesTypes => [CustomPaint];
+
+  @override
   CaptureNodeSemantics? captureSemantics(
     Element element,
     CapturedViewAttributes attributes,
-    CapturePrivacy capturePrivacy,
+    TreeCapturePrivacy capturePrivacy,
   ) {
     final widget = element.widget;
     if (widget is! CustomPaint) return null;