+++ packages/datadog_session_replay/test/capture/recorder_test.dart
@@ -25,7 +25,10 @@ class MockElement extends Mock implements Element {
   }
 }
 
-class MockElementRecorder extends Mock implements ElementRecorder {}
+class MockElementRecorder extends Mock implements ElementRecorder {
+  @override
+  List<Type> get handlesTypes => [SimpleTestCapture, Placeholder, Center];
+}
 
 class MockCaptureNode extends Mock implements CaptureNode {}
 
@@ -45,10 +48,12 @@ void main() {
       when(() => mockTimeProvider.now()).thenReturn(expectedDateTime);
       recorder = SessionReplayRecorder(
         timeProvider: mockTimeProvider,
-        defaultCapturePrivacy: CapturePrivacy(
+        defaultCapturePrivacy: TreeCapturePrivacy(
           textAndInputPrivacyLevel:
               TextAndInputPrivacyLevel.maskSensitiveInputs,
+          imagePrivacyLevel: ImagePrivacyLevel.maskNonAssetsOnly,
         ),
+        touchPrivacyLevel: TouchPrivacyLevel.show,
       );
     });
 
@@ -83,10 +88,12 @@ void main() {
       recorder = SessionReplayRecorder.withCustomRecorders(
         [mockRecorderA],
         timeProvider: mockTimeProvider,
-        defaultCapturePrivacy: CapturePrivacy(
+        defaultCapturePrivacy: TreeCapturePrivacy(
           textAndInputPrivacyLevel:
               TextAndInputPrivacyLevel.maskSensitiveInputs,
+          imagePrivacyLevel: ImagePrivacyLevel.maskNonAssetsOnly,
         ),
+        touchPrivacyLevel: TouchPrivacyLevel.show,
       );
 
       registerFallbackValue(
@@ -97,8 +104,10 @@ void main() {
         ),
       );
       registerFallbackValue(
-        CapturePrivacy(
-          textAndInputPrivacyLevel: TextAndInputPrivacyLevel.maskAll,
+        TreeCapturePrivacy(
+          textAndInputPrivacyLevel:
+              TextAndInputPrivacyLevel.maskSensitiveInputs,
+          imagePrivacyLevel: ImagePrivacyLevel.maskNonAssetsOnly,
         ),
       );
       registerFallbackValue(MockElement());
@@ -189,7 +198,6 @@ void main() {
         final testedTree = SimpleTestCapture(
           key: UniqueKey(),
           recorder: recorder,
-          child: Container(),
         );
         await tester.pumpWidget(testedTree);
         final capture = await recorder.performCapture();
@@ -238,7 +246,7 @@ void main() {
       expect(capture!.viewTreeSnapshot.nodes.length, 3);
     });
 
-    testWidgets('cpature subtree passes new capture privacy when overwritten', (
+    testWidgets('capture subtree passes new capture privacy when overwritten', (
       tester,
     ) async {
       // Given
@@ -252,8 +260,9 @@ void main() {
         }
         return SpecificElement(
           subtreeStrategy: subtreeStrategy,
-          subtreePrivacy: CapturePrivacy(
+          subtreePrivacy: TreeCapturePrivacy(
             textAndInputPrivacyLevel: TextAndInputPrivacyLevel.maskAll,
+            imagePrivacyLevel: ImagePrivacyLevel.maskAll,
           ),
           nodes: [MockCaptureNode()],
         );
@@ -278,16 +287,18 @@ void main() {
         () => mockRecorderA.captureSemantics(
           any(),
           any(),
-          CapturePrivacy(
+          TreeCapturePrivacy(
             textAndInputPrivacyLevel:
                 TextAndInputPrivacyLevel.maskSensitiveInputs,
+            imagePrivacyLevel: ImagePrivacyLevel.maskNonAssetsOnly,
           ),
         ),
         () => mockRecorderA.captureSemantics(
           any(),
           any(),
-          CapturePrivacy(
+          TreeCapturePrivacy(
             textAndInputPrivacyLevel: TextAndInputPrivacyLevel.maskAll,
+            imagePrivacyLevel: ImagePrivacyLevel.maskAll,
           ),
         ),
       ]);
@@ -426,10 +437,12 @@ void main() {
       );
       recorder = SessionReplayRecorder.withCustomRecorders(
         [mockElementRecorder],
-        defaultCapturePrivacy: CapturePrivacy(
+        defaultCapturePrivacy: TreeCapturePrivacy(
           textAndInputPrivacyLevel:
               TextAndInputPrivacyLevel.maskSensitiveInputs,
+          imagePrivacyLevel: ImagePrivacyLevel.maskNone,
         ),
+        touchPrivacyLevel: TouchPrivacyLevel.show,
       );
     });
 
@@ -468,7 +481,7 @@ void main() {
         key: UniqueKey(),
         recorder: recorder,
         child: PointerRecorder(
-          snapshotRecorder: mockPointerRecorder,
+          pointerRecorder: mockPointerRecorder,
           child: Placeholder(),
         ),
       );